requirements.txt
python-consul
jinja2
docopt
